Several travis issues fixed
